add giac configure option/check, add giac mpoly file
