+++ ginac/mpoly-giac.cpp
@@ -1,6 +1,6 @@
-/** @file normal.h
+/** @file normal.cpp
  *
- *  This file defines several functions that work on univariate and
+ *  This file implements several functions that work on univariate and
  *  multivariate polynomials and rational functions.
  *  These functions include polynomial quotient and remainder, GCD and LCM
  *  computation, square-free factorization and rational function normalization. */
@@ -23,20 +23,45 @@
  *  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-#ifndef __GINAC_NORMAL_H__
-#define __GINAC_NORMAL_H__
+#ifdef HAVE_GIAC
 
+#include "normal.h"
 #include "basic.h"
+#include "ex.h"
+#include "add.h"
+#include "constant.h"
+#include "expairseq.h"
+#include "mul.h"
+#include "numeric.h"
+#include "power.h"
+#include "operators.h"
+#include "pseries.h"
+#include "symbol.h"
+
+#include "giac.h"
 
 namespace GiNaC {
 
-class ex;
-class symbol;
+numeric lcm_of_coefficients_denominators(const ex &e)
+{
+        return *_num0_p;
+}
 
-extern bool get_first_symbol(const ex &e, ex &x);
-extern ex collect_common_factors(const ex & e);
-extern ex replace_with_symbol(const ex & e, exmap & repl);
+ex find_common_factor(const ex & e, ex & factor, exmap & repl)
+{
+        return _ex0;
+}
+
+ex gcd(const ex &a, const ex &b, ex *ca=nullptr, ex *cb=nullptr, bool check_args=true)
+{
+        return _ex0;
+}
+
+ex multiply_lcm(const ex &e, const numeric &lcm)
+{
+        return _ex0;
+}
 
 } // namespace GiNaC
 
-#endif // ndef __GINAC_NORMAL_H__
+#endif // HAVE_GIAC
