+++ ginac/normal.cpp
@@ -182,7 +182,7 @@ static ex replace_with_symbol(const ex & e, exmap & repl, exmap & rev_lookup)
  *  symbol is returned.
  *  @see basic::to_rational
  *  @see basic::to_polynomial */
-static ex replace_with_symbol(const ex & e, exmap & repl)
+ex replace_with_symbol(const ex & e, exmap & repl)
 {
 	// Since the repl contains replaced expressions we should search for them
 	ex e_replaced = e.subs(repl, subs_options::no_pattern);
@@ -572,198 +572,6 @@ ex ex::numer_denom() const
 		return e.subs(repl, subs_options::no_pattern);
 }
 
-
-/** Rationalization of non-rational functions.
- *  This function converts a general expression to a rational function
- *  by replacing all non-rational subexpressions (like non-rational numbers,
- *  non-integer powers or functions like sin(), cos() etc.) to temporary
- *  symbols. This makes it possible to use functions like gcd() and divide()
- *  on non-rational functions by applying to_rational() on the arguments,
- *  calling the desired function and re-substituting the temporary symbols
- *  in the result. To make the last step possible, all temporary symbols and
- *  their associated expressions are collected in the map specified by the
- *  repl parameter, ready to be passed as an argument to ex::subs().
- *
- *  @param repl collects all temporary symbols and their replacements
- *  @return rationalized expression */
-ex ex::to_rational(exmap & repl) const
-{
-	return bp->to_rational(repl);
-}
-
-// GiNaC 1.1 compatibility function
-ex ex::to_rational(lst & repl_lst) const
-{
-	// Convert lst to exmap
-	exmap m;
-	for (const auto & elem : repl_lst)
-		m.insert(std::make_pair(elem.op(0), elem.op(1)));
-
-	ex ret = bp->to_rational(m);
-
-	// Convert exmap back to lst
-	repl_lst.remove_all();
-	for (const auto& elem : m)
-		repl_lst.append(elem.first == elem.second);
-
-	return ret;
-}
-
-ex ex::to_polynomial(exmap & repl) const
-{
-	return bp->to_polynomial(repl);
-}
-
-// GiNaC 1.1 compatibility function
-ex ex::to_polynomial(lst & repl_lst) const
-{
-	// Convert lst to exmap
-	exmap m;
-	for (const auto & elem : repl_lst)
-		m.insert(std::make_pair(elem.op(0), elem.op(1)));
-
-	ex ret = bp->to_polynomial(m);
-
-	// Convert exmap back to lst
-	repl_lst.remove_all();
-	for (const auto& elem : m)
-		repl_lst.append(elem.first == elem.second);
-
-	return ret;
-}
-
-/** Default implementation of ex::to_rational(). This replaces the object with
- *  a temporary symbol. */
-ex basic::to_rational(exmap & repl) const
-{
-	return replace_with_symbol(*this, repl);
-}
-
-ex basic::to_polynomial(exmap & repl) const
-{
-	return replace_with_symbol(*this, repl);
-}
-
-
-/** Implementation of ex::to_rational() for symbols. This returns the
- *  unmodified symbol. */
-ex symbol::to_rational(exmap & repl) const
-{
-	return *this;
-}
-
-/** Implementation of ex::to_polynomial() for symbols. This returns the
- *  unmodified symbol. */
-ex symbol::to_polynomial(exmap & repl) const
-{
-	return *this;
-}
-
-
-/** Implementation of ex::to_rational() for a numeric. It splits complex
- *  numbers into re+I*im and replaces I and non-rational real numbers with a
- *  temporary symbol. */
-ex numeric::to_rational(exmap & repl) const
-{
-	if (is_real()) {
-		if (!is_rational())
-			return replace_with_symbol(*this, repl);
-	} else { // complex
-		numeric re = real();
-		numeric im = imag();
-		ex re_ex = re.is_rational() ? re : replace_with_symbol(re, repl);
-		ex im_ex = im.is_rational() ? im : replace_with_symbol(im, repl);
-		return re_ex + im_ex * replace_with_symbol(I, repl);
-	}
-	return *this;
-}
-
-/** Implementation of ex::to_polynomial() for a numeric. It splits complex
- *  numbers into re+I*im and replaces I and non-integer real numbers with a
- *  temporary symbol. */
-ex numeric::to_polynomial(exmap & repl) const
-{
-	if (is_real()) {
-		if (!is_integer())
-			return replace_with_symbol(*this, repl);
-	} else { // complex
-		numeric re = real();
-		numeric im = imag();
-		ex re_ex = re.is_integer() ? re : replace_with_symbol(re, repl);
-		ex im_ex = im.is_integer() ? im : replace_with_symbol(im, repl);
-		return re_ex + im_ex * replace_with_symbol(I, repl);
-	}
-	return *this;
-}
-
-
-/** Implementation of ex::to_rational() for powers. It replaces non-integer
- *  powers by temporary symbols. */
-ex power::to_rational(exmap & repl) const
-{
-	if (exponent.info(info_flags::integer))
-		return power(basis.to_rational(repl), exponent);
-	else
-		return replace_with_symbol(*this, repl);
-}
-
-/** Implementation of ex::to_polynomial() for powers. It replaces non-posint
- *  powers by temporary symbols. */
-ex power::to_polynomial(exmap & repl) const
-{
-	if (exponent.info(info_flags::posint))
-		return power(basis.to_rational(repl), exponent);
-	else if (exponent.info(info_flags::negint))
-	{
-		ex basis_pref = collect_common_factors(basis);
-		if (is_exactly_a<mul>(basis_pref) || is_exactly_a<power>(basis_pref)) {
-			// (A*B)^n will be automagically transformed to A^n*B^n
-			ex t = power(basis_pref, exponent);
-			return t.to_polynomial(repl);
-		}
-		else
-			return power(replace_with_symbol(power(basis, _ex_1), repl), -exponent);
-	} 
-	else
-		return replace_with_symbol(*this, repl);
-}
-
-
-/** Implementation of ex::to_rational() for expairseqs. */
-ex expairseq::to_rational(exmap & repl) const
-{
-	epvector s;
-	s.reserve(seq.size());
-	auto i = seq.begin(), end = seq.end();
-	while (i != end) {
-		s.push_back(split_ex_to_pair(recombine_pair_to_ex(*i).to_rational(repl)));
-		++i;
-	}
-	ex oc = overall_coeff.to_rational(repl);
-	if (oc.info(info_flags::numeric))
-		return thisexpairseq(s, overall_coeff);
-	else
-		s.push_back(combine_ex_with_coeff_to_pair(oc, _ex1));
-	return thisexpairseq(s, default_overall_coeff());
-}
-
-/** Implementation of ex::to_polynomial() for expairseqs. */
-ex expairseq::to_polynomial(exmap & repl) const
-{
-	epvector s;
-	s.reserve(seq.size());
-	auto i = seq.begin(), end = seq.end();
-	while (i != end) {
-		s.push_back(split_ex_to_pair(recombine_pair_to_ex(*i).to_polynomial(repl)));
-		++i;
-	}
-	ex oc = overall_coeff.to_polynomial(repl);
-	if (oc.info(info_flags::numeric))
-		return thisexpairseq(s, overall_coeff);
-	else
-		s.push_back(combine_ex_with_coeff_to_pair(oc, _ex1));
-	return thisexpairseq(s, default_overall_coeff());
-}
 
 
 } // namespace GiNaC
